+++ index.html
@@ -10,14 +10,6 @@
 
 <main class="central"></main>
 
-<template id="intro"></template>
-<template id="greeting"></template>
-<template id="rules"></template>
-<template id="game-1"></template>
-<template id="game-2"></template>
-<template id="game-3"></template>
-<template id="stats"></template>
-
 <script src="js/main.js"></script>
 </body>
 </html>
