+++ js/frame-change.js
@@ -1,7 +1,7 @@
 const mainFrame = document.querySelector(`main.central`);
-const frameChange = (getFrameValue) => {
+const frameChange = (frameTemplate) => {
   mainFrame.innerHTML = ``;
-  mainFrame.appendChild(getFrameValue);
+  mainFrame.appendChild(frameTemplate);
 };
 
 export default frameChange;
