+++ js/frame-create.js
@@ -1,0 +1,9 @@
+const frameCreate = (frameString) => {
+  const frameContainer = document.createElement(`div`);
+
+  frameContainer.innerHTML = frameString;
+
+  return frameContainer;
+};
+
+export default frameCreate;
