+++ js/third-game-frame.js
@@ -1,8 +1,8 @@
-import frameCreateFunc from './frameCreator.js';
-import frameChange from './frameChanger.js';
-import secondGame from './secondGameFrame.js';
+import frameCreate from './frame-create';
+import frameChange from './frame-change';
+import stats from './stats-frame';
 
-const firstGameTemplate = `<header class="header">
+const thirdGameTemplate = `<header class="header">
                             <div class="header__back">
                               <button class="back">
                                 <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
@@ -17,29 +17,16 @@ const firstGameTemplate = `<header class="header">
                             </div>
                           </header>
                           <div class="game">
-                            <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
-                            <form class="game__content">
+                            <p class="game__task">Найдите рисунок среди изображений</p>
+                            <form class="game__content  game__content--triple">
                               <div class="game__option">
-                                <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
-                                <label class="game__answer game__answer--photo">
-                                  <input name="question1" type="radio" value="photo">
-                                  <span>Фото</span>
-                                </label>
-                                <label class="game__answer game__answer--paint">
-                                  <input name="question1" type="radio" value="paint">
-                                  <span>Рисунок</span>
-                                </label>
+                                <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+                              </div>
+                              <div class="game__option  game__option--selected">
+                                <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
                               </div>
                               <div class="game__option">
-                                <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
-                                <label class="game__answer  game__answer--photo">
-                                  <input name="question2" type="radio" value="photo">
-                                  <span>Фото</span>
-                                </label>
-                                <label class="game__answer  game__answer--paint">
-                                  <input name="question2" type="radio" value="paint">
-                                  <span>Рисунок</span>
-                                </label>
+                                <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
                               </div>
                             </form>
                             <div class="stats">
@@ -48,11 +35,11 @@ const firstGameTemplate = `<header class="header">
                                 <li class="stats__result stats__result--slow"></li>
                                 <li class="stats__result stats__result--fast"></li>
                                 <li class="stats__result stats__result--correct"></li>
+                                <li class="stats__result stats__result--wrong"></li>
                                 <li class="stats__result stats__result--unknown"></li>
+                                <li class="stats__result stats__result--slow"></li>
                                 <li class="stats__result stats__result--unknown"></li>
-                                <li class="stats__result stats__result--unknown"></li>
-                                <li class="stats__result stats__result--unknown"></li>
-                                <li class="stats__result stats__result--unknown"></li>
+                                <li class="stats__result stats__result--fast"></li>
                                 <li class="stats__result stats__result--unknown"></li>
                               </ul>
                             </div>
@@ -68,24 +55,14 @@ const firstGameTemplate = `<header class="header">
                             </div>
                           </footer>`;
 
-const firstGame = frameCreateFunc(firstGameTemplate, `game-1`);
+const thirdGame = frameCreate(thirdGameTemplate);
 
-const answerHandler = () => {
-  const answers = firstGame.querySelectorAll(`.game__answer input`);
-  let answerCounter = 0;
-
-  if (event.target.closest(`.game__answer`)) {
-    answers.forEach((answer) => {
-      if (answer.checked) {
-        answerCounter++;
-        if (answerCounter === 2) {
-          frameChange(secondGame);
-        }
-      }
-    });
+const answerHandler = (event) => {
+  if (event.target.closest(`.game__option`)) {
+    frameChange(stats);
   }
 };
 
-firstGame.addEventListener(`change`, answerHandler);
+thirdGame.addEventListener(`click`, answerHandler);
 
-export default firstGame;
+export default thirdGame;
